# contracts/openapi.yaml
components:
  schemas:
    ErrorEnvelope:
      properties:
        data:
          type: "null"
        error:
          properties:
            code:
              type: string
            message:
              type: string
          required:
            - code
            - message
          type: object
      required:
        - data
        - error
      type: object
  securitySchemes:
    ingestBearer:
      description: Dedicated INGEST_API_TOKEN
      scheme: bearer
      type: http
    sessionBearer:
      description: Signed session returned by auth.unlock
      scheme: bearer
      type: http
info:
  description: Private pre-1.0 API. Backward compatibility is not guaranteed; frontend and backend change together.
  title: Wallet Web Personal API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/cards/{id}/summary:
    get:
      operationId: cards.summary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $schema: https://json-schema.org/draft/2020-12/schema
                  error:
                    type: "null"
                required:
                  - data
                  - error
                type: object
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error 400
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error 401
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error 404
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error 500
      security:
        - sessionBearer: []
      summary: Get credit-card cycle summary
      tags:
        - cards
      x-stability: stable
...
  /api/wallet/bootstrap:
    post:
      operationId: wallet.bootstrap
      requestBody:
        content:
          application/json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              additionalProperties: false
              properties:
                recordsCursor:
                  anyOf:
                    - maxLength: 500
                      minLength: 1
                      type: string
                    - type: "null"
                  default: null
                recordsLimit:
                  default: 200
                  maximum: 500
                  minimum: 1
                  type: integer
              required:
                - recordsLimit
                - recordsCursor
              type: object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                properties:
                  data:
                    $schema: https://json-schema.org/draft/2020-12/schema
                  error:
                    type: "null"
                required:
                  - data
                  - error
                type: object
          description: Success
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error 400
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error 401
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
          description: Error 500
      security:
        - sessionBearer: []
      summary: Generate recurring debts and load a paginated snapshot
      tags:
        - wallet
      x-stability: stable
servers:
  - description: Current deployment
    url: /